hugginhface/pyproject.toml
[tool.poetry]
name = "hugginhface"
version = "0.1.0"
description = ""
authors = ["year <sainknight87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
transformers = "^4.45.1"
torch = {version = "^2.4.1+cu121", source = "pytorch"}
torchvision = {version = "^0.19.1+cu121", source = "pytorch"}
torchaudio = {version = "^2.4.1+cu121", source = "pytorch"}
python-dotenv = "^1.0.1"
nltk = "^3.9.1"
openai = "^1.51.0"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
